+++ src/egon/data/datasets/hydrogen_etrago/__init__.py
@@ -1,10 +1,11 @@
 """The central module containing all code dealing with heat sector in etrago
 """
 from egon.data.datasets import Dataset
-from egon.data.datasets.hydrogen_etrago.bus import insert_hydrogen_buses
+from egon.data.datasets.hydrogen_etrago.bus import insert_h2_buses_eGon100RE, insert_hydrogen_buses
+from egon.data.datasets.hydrogen_etrago.h2_grid import insert_h2_pipelines
 from egon.data.datasets.hydrogen_etrago.h2_to_ch4 import insert_h2_to_ch4_to_h2
 from egon.data.datasets.hydrogen_etrago.power_to_h2 import (
-    insert_power_to_h2_to_power,
+    insert_power_to_h2_to_power, insert_power_to_h2_to_power_eGon100RE
 )
 from egon.data.datasets.hydrogen_etrago.storage import (
     calculate_and_map_saltcavern_storage_potential,
@@ -46,6 +47,18 @@ class HydrogenPowerLinkEtrago(Dataset):
         )
 
 
+class HydrogenPowerLinkEtragoeGon100RE(Dataset):
+    def __init__(self, dependencies):
+        super().__init__(
+            name="HydrogenPowerLinkEtragoeGon100RE",
+            version="0.0.0",
+            dependencies=dependencies,
+            tasks=(
+                insert_power_to_h2_to_power_eGon100RE
+            ),
+        )
+
+
 class HydrogenMethaneLinkEtrago(Dataset):
     def __init__(self, dependencies):
         super().__init__(
@@ -53,4 +66,17 @@ class HydrogenMethaneLinkEtrago(Dataset):
             version="0.0.0",
             dependencies=dependencies,
             tasks=(insert_h2_to_ch4_to_h2),
+        )
+
+
+class HydrogenGridEtrago(Dataset):
+    def __init__(self, dependencies):
+        super().__init__(
+            name="HydrogenGridEtrago",
+            version="0.0.0",
+            dependencies=dependencies,
+            tasks=(
+                insert_h2_buses_eGon100RE,
+                insert_h2_pipelines,
+            ),
         )