+++ src/egon/data/datasets/industrial_gas_demand.py
@@ -11,13 +11,13 @@ import requests
 
 from egon.data import db
 from egon.data.config import settings
-from egon.data.datasets import Dataset, scenario_parameters
-from egon.data.datasets.etrago_setup import link_geom_from_buses
-from egon.data.datasets.gas_prod import assign_bus_id
-from egon.data.datasets.insert_etrago_buses import (
+from egon.data.datasets import Dataset
+from egon.data.datasets.etrago_helpers import (
     finalize_bus_insertion,
     initialise_bus_insertion,
 )
+from egon.data.datasets.etrago_setup import link_geom_from_buses
+from egon.data.datasets.gas_prod import assign_bus_id
 from egon.data.datasets.scenario_parameters import get_sector_parameters